+++ Scripts/SmallEyeItem.cs
@@ -6,26 +6,39 @@ public class SmallEyeItem: NoisemakerProp
 {
     public Animator Animator;
     private static readonly int Angry = Animator.StringToHash("Angry");
+    public AudioClip grabSFX;
 
     public override void Start()
     {
         base.Start();
-        noiseAudio.volume = BigEyesPlugin.instance.smallEyesScrapVolume.Value;
-        noiseAudioFar.volume = BigEyesPlugin.instance.smallEyesScrapVolume.Value;
+        ResetVolume();
     }
 
     public override void PocketItem()
     {
         base.PocketItem();
-        noiseAudio.volume = BigEyesPlugin.instance.smallEyesScrapVolume.Value;
-        noiseAudioFar.volume = BigEyesPlugin.instance.smallEyesScrapVolume.Value;
+        ResetVolume();
     }
 
     public override void ItemActivate(bool used, bool buttonDown = true)
     {
-        noiseAudio.volume = BigEyesPlugin.instance.smallEyesScrapVolume.Value;
-        noiseAudioFar.volume = BigEyesPlugin.instance.smallEyesScrapVolume.Value;
+        ResetVolume();
         base.ItemActivate(used, buttonDown);
         Animator.SetTrigger(Angry);
+    }
+
+    public void ResetVolume()
+    {
+        noiseAudio.volume = BigEyesPlugin.instance.smallEyesScrapVolume.Value;
+        noiseAudioFar.volume = BigEyesPlugin.instance.smallEyesScrapVolume.Value;
+        minLoudness = BigEyesPlugin.instance.smallEyesScrapVolume.Value;
+        maxLoudness = BigEyesPlugin.instance.smallEyesScrapVolume.Value;
+    }
+
+    public override void GrabItem()
+    {
+        ResetVolume();
+        noiseAudio.PlayOneShot(grabSFX);
+        base.GrabItem();
     }
 }